會員管理
更新後台會員管理 可以從資料庫顯示
更動redisterDao的屬性<RegisterModel, Integer>

+++ workspace/springbootdemo/src/main/java/com/loveGod/demo/model/RegisterDao.java
@@ -1,16 +1,19 @@
 package com.loveGod.demo.model;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 //import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface RegisterDao extends JpaRepository<RegisterModel, Long> {
+public interface RegisterDao extends JpaRepository<RegisterModel, Integer> {
 
 	//================================== 使用者登入判斷帳密是否正確 ========================
 	@Query("select r from RegisterModel r where r.memberId = ?1 and r.password = ?2 ")
 	List<RegisterModel> findLogin(String memberId, String password);
+
+
